move useCurrentKibanaTheme and useCurrentEuiTheme
