+++ x-pack/plugins/ml/public/application/components/color_range_legend/index.ts
@@ -13,5 +13,4 @@ export {
   useColorRange,
   COLOR_RANGE,
   COLOR_RANGE_SCALE,
-  useCurrentEuiTheme,
 } from './use_color_range';
