+++ x-pack/plugins/ml/public/application/explorer/swimlane_annotation_container.tsx
@@ -10,9 +10,9 @@ import d3 from 'd3';
 import { scaleTime } from 'd3-scale';
 import { i18n } from '@kbn/i18n';
 import moment from 'moment';
+import { useCurrentEuiTheme } from '../contexts/kibana';
 import type { Annotation, AnnotationsTable } from '../../../common/types/annotations';
 import { ChartTooltipService } from '../components/chart_tooltip';
-import { useCurrentEuiTheme } from '../components/color_range_legend';
 
 export const Y_AXIS_LABEL_WIDTH = 170;
 export const Y_AXIS_LABEL_PADDING = 8;
