+++ x-pack/plugins/ml/public/application/memory_usage/memory_tree_map/tree_map.tsx
@@ -19,7 +19,7 @@ import { FIELD_FORMAT_IDS } from '@kbn/field-formats-plugin/common';
 import { EuiComboBox, EuiComboBoxOptionOption, EuiEmptyPrompt, EuiSpacer } from '@elastic/eui';
 import { i18n } from '@kbn/i18n';
 import { FormattedMessage } from '@kbn/i18n-react';
-import { useCurrentKibanaTheme } from '../../components/color_range_legend/use_color_range';
+import { useCurrentKibanaTheme } from '../../contexts/kibana/use_current_theme';
 import { MemoryUsageInfo } from '../../../../common/types/trained_models';
 import { JobType, MlSavedObjectType } from '../../../../common/types/saved_objects';
 import { useTrainedModelsApiService } from '../../services/ml_api_service/trained_models';
