+++ x-pack/plugins/ml/public/application/model_management/test_models/models/question_answering/question_answering_output.tsx
@@ -10,8 +10,7 @@ import useObservable from 'react-use/lib/useObservable';
 
 import { EuiBadge, EuiHorizontalRule } from '@elastic/eui';
 
-import { useCurrentEuiTheme } from '../../../../components/color_range_legend/use_color_range';
-
+import { useCurrentEuiTheme } from '../../../../contexts/kibana';
 import type {
   QuestionAnsweringInference,
   FormattedQuestionAnsweringResult,
